Change the handler to the right place Add AsyncStorage to store the last notification received Add some logic to show the last notification in AsyncStorage to the user
