+++ sample/package.json
@@ -10,6 +10,7 @@
     "lint": "eslint ."
   },
   "dependencies": {
+    "@react-native-async-storage/async-storage": "^1.13.2",
     "react": "16.9.0",
     "react-native": "0.61.2",
     "react-native-android-notification-listener": "../"
@@ -17,24 +18,24 @@
   "devDependencies": {
     "@babel/core": "^7.6.2",
     "@babel/runtime": "^7.6.2",
-    "babel-jest": "^24.9.0",
-    "jest": "^24.9.0",
-    "metro-react-native-babel-preset": "^0.56.0",
-    "react-test-renderer": "16.9.0",
-    "eslint": "^6.8.0",
+    "@react-native-community/eslint-config": "^1.1.0",
     "@types/react-native": "^0.62.18",
     "@types/react-redux": "^7.1.9",
+    "babel-jest": "^24.9.0",
+    "eslint": "^6.8.0",
     "eslint-config-airbnb": "^18.0.1",
     "eslint-config-defaults": "^9.0.0",
     "eslint-config-prettier": "^6.9.0",
     "eslint-import-resolver-node": "^0.3.3",
-    "@react-native-community/eslint-config": "^1.1.0",
     "eslint-plugin-import": "^2.18.2",
     "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-prettier": "^3.1.2",
     "eslint-plugin-react": "^7.14.3",
     "eslint-plugin-react-hooks": "^1.7.0",
-    "eslint-plugin-prettier": "^3.1.2",
-    "prettier": "^1.19.1"
+    "jest": "^24.9.0",
+    "metro-react-native-babel-preset": "^0.56.0",
+    "prettier": "^1.19.1",
+    "react-test-renderer": "16.9.0"
   },
   "jest": {
     "preset": "react-native"
